test: integ tests for ESLint initial
