+++ test/spec/ESLintExtensionTest-files/ESLint_JSHint/error.js
@@ -1,0 +1,2 @@
+console.log("hello")
+const x;
