+++ test/spec/ESLintExtensionTest-files/ESLint_JSHint/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "misc",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "type": "module",
+  "scripts": {},
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "eslint": "6.0.0"
+  }
+}
