+++ test/spec/ESLintExtensionTest-files/es6/error.js
@@ -1,0 +1,2 @@
+console.log("hello")
+const x;
